# backend/docs/swagger.yaml
definitions:
  entity.ReminderDto:
    properties:
      before:
        items:
          type: integer
        type: array
      id:
        type: integer
      msg:
        type: string
      when:
        type: string
    required:
    - before
    - id
    - msg
    - when
    type: object
  entity.UserDto:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
info:
  contact: {}
  description: This is a sample swagger for CK project API
  termsOfService: http://swagger.io/terms/
  title: CK project API
  version: "1.0"
paths:
  /aunt/addremind:
    post:
      consumes:
      - application/json
      description: AddReminders
      parameters:
      - description: jwt token
        in: header
        name: token
        required: true
        type: string
      - description: Request of Creating User Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.ReminderDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: AddReminders
      tags:
      - Reminders
    put:
      consumes:
      - application/json
      parameters:
      - description: jwt token
        in: header
        name: token
        required: true
        type: string
      - description: Request of Creating User Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.ReminderDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      tags:
      - Reminders
  /aunt/user:
    delete:
      consumes:
      - application/json
      parameters:
      - description: jwt token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      tags:
      - Reminders
    get:
      consumes:
      - application/json
      parameters:
      - description: jwt token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Getting Reminds by token
      tags:
      - Reminders
  /login:
    get:
      consumes:
      - application/json
      description: Login User by jso in detail
      parameters:
      - description: Login
        in: header
        name: login
        required: true
        type: string
      - description: Password
        in: header
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Login User by json
      tags:
      - Users
  /user:
    post:
      consumes:
      - application/json
      description: Add User by jso in detail
      parameters:
      - description: Request of Creating User Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Add User by json
      tags:
      - Users
swagger: "2.0"